# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TinyClaw API",
    "version": "1",
    "description": "HTTP API for the TinyClaw personal AI assistant."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4310",
      "description": "Local dev server"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "ModelsResponse": {
        "type": "object",
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {},
              "additionalProperties": {}
            }
          }
        },
        "required": [
          "models"
        ],
        "additionalProperties": {}
      },
      "DiscoverModelsRequest": {
        "type": "object",
        "properties": {
          "baseUrl": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          }
        },
        "required": [
          "baseUrl"
        ]
      },
      "ListProvidersResponse": {
        "type": "object",
        "properties": {
          "providers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {},
              "additionalProperties": {}
            }
          }
        },
        "required": [
          "providers"
        ],
        "additionalProperties": {}
      },
      "CreateProviderResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "CreateProviderRequest": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "UpdateProviderResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "UpdateProviderRequest": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "DeleteProviderResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "ConfigureProviderResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "ConfigureProviderRequest": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "ListTimezonesResponse": {
        "type": "object",
        "properties": {
          "timezones": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {},
              "additionalProperties": {}
            }
          }
        },
        "required": [
          "timezones"
        ],
        "additionalProperties": {}
      },
...
    },
...
  },
  "paths": {
...
    "/v1/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List available models",
        "operationId": "listModels",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "catalog",
                "remote"
              ]
            },
            "required": false,
            "name": "source",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Model catalog",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models/discover": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Discover models from a provider base URL",
        "operationId": "discoverModels",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverModelsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model catalog",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/providers": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List configured provider instances",
        "operationId": "listProviders",
        "responses": {
          "200": {
            "description": "Provider instances",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProvidersResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Add a provider instance",
        "operationId": "createProvider",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProviderResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/providers/{providerId}": {
      "patch": {
        "tags": [
          "Models"
        ],
        "summary": "Update a provider instance",
        "operationId": "updateProvider",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "providerId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProviderResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Models"
        ],
        "summary": "Remove a provider instance",
        "operationId": "deleteProvider",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "providerId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProviderResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/provider": {
      "put": {
        "tags": [
          "Models"
        ],
        "summary": "Configure the LLM provider and API key",
        "operationId": "configureProvider",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureProviderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigureProviderResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/timezones": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List available timezones",
        "operationId": "listTimezones",
        "responses": {
          "200": {
            "description": "Timezone catalog",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTimezonesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/timezone": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get the user timezone",
        "operationId": "getTimezone",
        "responses": {
          "200": {
            "description": "Timezone settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimezoneSettingsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Models"
        ],
        "summary": "Update the user timezone",
        "operationId": "setTimezone",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTimezoneRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Timezone settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimezoneSettingsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/thinking": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get thinking settings",
        "operationId": "getThinkingSettings",
        "responses": {
          "200": {
            "description": "Thinking settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThinkingSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Models"
        ],
        "summary": "Update thinking settings",
        "operationId": "setThinkingSettings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThinkingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thinking settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThinkingSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/telegram": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get Telegram settings",
        "operationId": "getTelegramSettings",
        "responses": {
          "200": {
            "description": "Telegram settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Models"
        ],
        "summary": "Update Telegram settings",
        "operationId": "setTelegramSettings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTelegramSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Telegram settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/telegram/handshake": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Regenerate Telegram handshake",
        "operationId": "regenerateTelegramHandshake",
        "responses": {
          "200": {
            "description": "Telegram settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/whatsapp": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get WhatsApp settings",
        "operationId": "getWhatsAppSettings",
        "responses": {
          "200": {
            "description": "WhatsApp settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatsAppSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Models"
        ],
        "summary": "Update WhatsApp settings",
        "operationId": "setWhatsAppSettings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWhatsAppSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "WhatsApp settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatsAppSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/whatsapp/pairing-code": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Regenerate WhatsApp pairing code",
        "operationId": "regenerateWhatsAppPairingCode",
        "responses": {
          "200": {
            "description": "WhatsApp settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatsAppSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/whatsapp/reconnect": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Reconnect WhatsApp session",
        "operationId": "reconnectWhatsApp",
        "responses": {
          "200": {
            "description": "WhatsApp settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatsAppSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}